Moved preview_content to floating_text and merged
